Allow properties in JSON to reference other properties
Use {some.property} to reference other properties within strings.
E.g. "page-title": "{heading} - Blogdown"

If file.name is given, it is lower cased and spaces are replaced with
dashes. Templates may contain a "file" property now as well. This can be
used to generate the file name out of a property in the JSON of the
actual pages.